Incorporated review feedback from @djmitche
